Match every width JS and the kernel disagreed on; the floor is libm
dd04380 admitted the parity claims were a size threshold: 0% at 256x256, 17-27%
at 768x768, because JS accumulated diffused error in f64 and the kernel in f32.
This closes that, and finds where it stops being closeable.

Three widths, all one shape — one side rounding where the other did not:

**The error arithmetic.** JS computed `er`, `scale`, `weight * scale` and
`er * weight` in f64; the kernel does each in f32. The store already agreed —
both write a Float32Array, and an f32+f32 sum computed in f64 and rounded once
is bit-equal to an f32 add — so only the products were wide. Math.fround'ed.
Nothing lost: the very next store discarded that precision anyway. Only the JS
fallback pays it, and a fround is one instruction once JIT'ed. The alternative,
widening the kernel's error buffer to f64, would cost 2x bandwidth on the path
that actually ships to fix a fallback nobody sees.

**The LUTs were not the same table.** Rust built its in f32 throughout, JS in
f64 rounded once at the store. 214/256 entries differed, worst 1.8e-7. Now
identical — verified 0/256 against a live export.

**RGB and RGB_APPROX scored distance in f32**, while HSV, LAB and all five
whole-buffer quantizers use f64, as does the JS colorDistance they mirror. They
were the only two still diverging once the arithmetic matched — 10% and 9% at
768x768 against 0% for the f64 three, which is what named them. Widened rather
than narrowing JS: colorDistance is shared with the whole-buffer fallback, whose
counterpart is f64 and already bit-parity-clean, so narrowing would fix this
path and break that one.

  Stucki 1024x1024, 16-colour, JS vs WASM, both linearize modes:
    RGB 0   RGB_APPROX 0   HSV 0   LAB 0   LEVELS 0
    OKLAB 0 (linearize=false) / 12.43% (linearize=true)

Eleven of twelve are bit-identical at 1024x1024, up from divergence at 12x9. A
property now, not a threshold: there is no last-bit difference left to amplify.

**The twelfth cannot be fixed this way, and that is the useful finding.**
Math.cbrt and Rust f64::cbrt are different libm implementations, disagreeing by
up to 1 ULP on 1032 of 12288 of the l/m/s values OKLab feeds them — 8.4%. cbrt
is not correctly-rounded under IEEE-754; both are entitled to their answers.
Closing it means implementing cbrt identically on both sides.

So OKLab's 0% at linearize=false is itself a threshold that has not fired, not a
guarantee, and Lab's powf is likely in the same position. Exact JS/WASM parity
for error diffusion is not achievable while both sides call their own
transcendentals. The systematic differences are gone; what is left is a floor.

I checked RGB's 27% at 768 against the stashed tree before believing it — 37.8%
before this change, so an improvement, not a regression. Worth the check: the
probe had also flipped serpentine on, and two variables moved at once.

059 rewritten to lead with the outcome and mark the superseded sections. It kept
its wrong turns on purpose — all three were real measurements read at a fixture
too small to show the phenomenon, which is the transferable part.

+++ packages/ditherer-filters/src/wasm/rgba2laba/src/lib.rs
@@ -827,13 +827,25 @@ fn palette_match_rgb(
             let qb = quant_levels_channel(sb, step);
             (qr, qg, qb, clamp_u8_f32(qr), clamp_u8_f32(qg), clamp_u8_f32(qb))
         }
+        // RGB and RGB_APPROX scored in f32 while every other distance in this
+        // crate — HSV and LAB below, and all five whole-buffer quantizers — uses
+        // f64, as does the JS `colorDistance` both are meant to mirror. They were
+        // the only two that diverged from the JS loop once the error arithmetic
+        // was matched: 10% and 9% of pixels at 768x768, against 0% for the f64
+        // three. A last-bit difference flips any pixel near the bisector between
+        // two entries, and error diffusion cascades it (docs/plan/059).
+        //
+        // Widened rather than frounding the JS side, because `colorDistance` is
+        // shared with the whole-buffer JS fallback, whose Rust counterpart is
+        // already f64 and bit-parity-clean. Narrowing JS would fix this path and
+        // break that one.
         PAL_MODE_RGB => {
             let mut best = 0usize;
-            let mut best_d = f32::MAX;
+            let mut best_d = f64::MAX;
             for (j, c) in pal_rgba.iter().enumerate() {
-                let dr = sr - c[0] as f32;
-                let dg = sg - c[1] as f32;
-                let db = sb - c[2] as f32;
+                let dr = sr as f64 - c[0] as f64;
+                let dg = sg as f64 - c[1] as f64;
+                let db = sb as f64 - c[2] as f64;
                 let d = dr*dr + dg*dg + db*db;
                 if d < best_d { best_d = d; best = j; }
             }
@@ -842,12 +854,12 @@ fn palette_match_rgb(
         }
         PAL_MODE_RGB_APPROX => {
             let mut best = 0usize;
-            let mut best_d = f32::MAX;
+            let mut best_d = f64::MAX;
             for (j, c) in pal_rgba.iter().enumerate() {
-                let rm = (sr + c[0] as f32) / 2.0;
-                let dr = sr - c[0] as f32;
-                let dg = sg - c[1] as f32;
-                let db = sb - c[2] as f32;
+                let rm = (sr as f64 + c[0] as f64) / 2.0;
+                let dr = sr as f64 - c[0] as f64;
+                let dg = sg as f64 - c[1] as f64;
+                let db = sb as f64 - c[2] as f64;
                 let d = (2.0 + rm / 256.0) * dr * dr
                     + 4.0 * dg * dg
                     + (2.0 + (255.0 - rm) / 256.0) * db * db;
